refactor: Move the FDv1 feature-store wrapper into its data system
